refactor: none-dead code , App Structure , run script phase warning , refactor key window
