--- Akla/Resources/HomeViewController.swift
@@ -1,28 +1,0 @@
-//
-//  HomeViewController.swift
-//  Akla
-//
-//  Created by Ahmad Ayman Mansour on 19/05/2023.
-//
-
-import UIKit
-
-class HomeViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        view.backgroundColor = .systemCyan
-    }
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
