+++ Akla/Resources/Localization/Localizable.strings
@@ -1,0 +1,1 @@
+'app.title' = "Akla";
