+++ Vendors/SwiftGen/bin/SwiftGen_SwiftGenCLI.bundle/Contents/Resources/templates/colors/literals-swift5.stencil
@@ -1,0 +1,47 @@
+// swiftlint:disable all
+// Generated using SwiftGen — https://github.com/SwiftGen/SwiftGen
+
+{% if palettes %}
+{% set enumName %}{{param.enumName|default:"ColorName"}}{% endset %}
+{% set accessModifier %}{% if param.publicAccess %}public{% else %}internal{% endif %}{% endset %}
+#if os(macOS)
+  import AppKit
+  {% if enumName != 'NSColor' %}
+  {{accessModifier}} enum {{enumName}} { }
+  {% endif %}
+#elseif os(iOS) || os(tvOS) || os(watchOS)
+  import UIKit
+  {% if enumName != 'UIColor' %}
+  {{accessModifier}} enum {{enumName}} { }
+  {% endif %}
+#endif
+
+// swiftlint:disable superfluous_disable_command
+// swiftlint:disable file_length
+
+// MARK: - Colors
+
+// swiftlint:disable identifier_name line_length type_body_length
+{{accessModifier}} extension {{enumName}} {
+{% macro h2f hex %}{{hex|hexToInt|int255toFloat}}{% endmacro %}
+{% macro enumBlock colors accessPrefix %}
+  {% for color in colors %}
+  /// 0x{{color.red}}{{color.green}}{{color.blue}}{{color.alpha}} (r: {{color.red|hexToInt}}, g: {{color.green|hexToInt}}, b: {{color.blue|hexToInt}}, a: {{color.alpha|hexToInt}})
+  {{accessPrefix}}static let {{color.name|swiftIdentifier:"pretty"|lowerFirstWord|escapeReservedKeywords}} = #colorLiteral(red: {%+ call h2f color.red %}, green: {%+ call h2f color.green %}, blue: {%+ call h2f color.blue %}, alpha: {%+ call h2f color.alpha %})
+  {% endfor %}
+{% endmacro %}
+  {% if palettes.count > 1 or param.forceFileNameEnum %}
+  {% set accessPrefix %}{{accessModifier}} {%+ endset %}
+  {% for palette in palettes %}
+  enum {{palette.name|swiftIdentifier:"pretty"|escapeReservedKeywords}} {
+    {% filter indent:2," ",true %}{% call enumBlock palette.colors accessPrefix %}{% endfilter %}
+  }
+  {% endfor %}
+  {% else %}
+  {% call enumBlock palettes.first.colors "" %}
+  {% endif %}
+}
+// swiftlint:enable identifier_name line_length type_body_length
+{% else %}
+// No color found
+{% endif %}
